+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 * Added an option to create case insensitive regular expressions.
 
+* Added a getter named `regExp` to `ParameterToken` for the regular expression
+  used to match arguments.
+
 ## 0.2.1
 
 * Removed unnecessary code.